Fix bug in csv_t::parseCsvRow
In previous commit. DATA, DATA,\0  These kind of data goes wrong.
The last element should be "", but iter points on the address of
'\0'. In this situation, the parsing task is finished.

In this commit, there is a flag, endline, to describe whether the
iter points on the end of line or not.