Add job_t and add the lot_t::job
lot_t::job is used to generate a job_t type variable from lot_t. The
information of generated job_t is all from lot_t. The information is
represented in job_info_t type that contains a union and 2 bit fields.
The union in job_info_t is the data field, the benefit of union is that
the program is able to compare two job_info_t variables by two arrays of
unsigned int variable rather then two arrays of character.