Add main.py and add functions to lots_t
In this version, main.py is added which is used to formalize the input
data. The formation includes removing the columns that is never used in
the program and summarize the input files in a single file named config.csv.

The functions added to lots_t is originally being in arrival.cpp The
lots_t object is in charge creating all lots object.