Fix comparison bug
In the previous version, I notice that the best fitness value isn't
decreasing stably. I find out the bug which is in the comparison
function passed into the qsort function. In previous version, comparison
function return the result of condition, but the result of condition is
literally 1 or 0 resulting in this bug. In this version, the return
value of comparison function is 1 if condition is true, otherwise,
return -1 if condition is false.