Input the algorithm parameters from config.json
In this version, the algorithm can read the parameters such as weight
and setup time from the config.json and that is convenient to run the
experiments.

The main change of this version is the way of computing setup time. In
previous version, the setup time in each setupTime function is a literal
constant instead of variable number. In this version, I add a structure
to link the relationship between setup time function and its setup
time, which is setup_time_unit_t. The last and imcomplete type field of
machine_base_operations_t is also edited.

In this version, I also add weights_t and scheduling_parameters_t these
two structure to store the weights and parameters. These two structures
are writen in include/parameters.h and which is included by
include/chromosome.h, include/machine.h and include/population.h.

I also add Makefile and some git hook scripts into the repo. The git
hook scripts include pre-commit hook and commit-msg hook. The function
of pre-commit hook is helping programmer check if the code follow the
.clang-format's rule. The function of commit-msg hook is helping
programmer write down good commit message. Makefile helps us install git
hooks easily.

Summary:
1. Change the way of computing setup time.
2. Link the relationship between setup time function and the number.
3. Add weights_t and scheduling_parameters_t to store the parameters.
4. Add Makefile and some git hooks.