Add the tool restriction
In the meeting in 2021/8/12, Kuka complainted on the processing rate of
the lots. And the point she mainly complaints about is why there are
so many lots whose uph is 0. I finally traced the code and I found the
reason is Kuka has told us that the part_no which starts with A0803
only pairs with the models UTC1000, UTC2000 and UTC3000 and the part_no
which starts with A0801 shouldn't pair with the models UTC1000, UTC2000
and UTC3000.

However, actually, the part_no which starts with A0803 can pair with the
tools UTC1000S, UTC2000S too but Kuka hasn't mentioned that before. She
only knows how to blame for that angrily rather then gives us correct
information.

Summary:
The mainly change for this commit is adding the pairable models for the
part_no A0803 and the pairable models are UTC1000, UTC1000S, UTC2000,
UTC2000S and UTC3000.