Fix duplicate uph data error
In previous version, if uph is zero, that kind of data will not
be considered. Actually, in the uph file, there are lots of rows
which are the same model but the value of uph are different.
And some of which entries, one entry's value is zero but the
other isn't. In this situation, the program goes wrong because
the code thinks that kind of model is no need to consider.

In this version, I use map to preprocess the data of uphs and if
the value is zero, it will not be store in map, and the container
accepts duplicate model.