Fix no wire error
Add wire_stock data to fix no wire error.
I found a stupid mistake in function lots_t::setAmountOfWires.
In that function I attempted to set different criteria for
different kind of wires. if part_id[4] is 'A' which is golden
wire otherwise, it is other kind of wire. I wrote down two
"== 'A'" in two different if-clauses so that the other kind of
wires is always insufficient. In this version, I fix it.