Finish stage2 scheduling algorithm
Stage2 scheduling algorithm has been well testing. I wrote the
algorithm last night and also wrote some test suits. Today, I
combine the code in main and run the algorithm, and I found that
there are some machines which has no job to run and the reason
is that the code to determine the job if it is qualified to run
on the machine is depending on the arrival time and the available
time of the machine. The ideal concept is let the machines have
no gap in the time line but that is resulting the empty machine.

In this version, I set the gap restriction to try to decrease the
number of empty machine but the affect is little. There still are
some empty machines QQ