Fix bug in timeConverter function
The function of timeConverter is converting the string input
from user to time_t number. The range of field tm_year
is the year since 1900, but I didn't notice that. Actually,
that results in integer overflow, but in Unix-like operating
system, that didn't affect the program result because we always
use the relative time to run the algorithm, and the output is
the relative time too. However, when we transplant the program
to Windows OS, we got the problem. The problem is that the
mktime function, used to transform the struct tm to time_t number,
in Windows, is not only converting the time since 1970/1/1 but
check the result is overflow or not. If the result is overflow, the
return value of mktime will be -1 so that the timeConverter does
affect the program's result.

In this version, I fix the problem.