Rewrite test suit for distributeAResource
In previous version, the test suit cause seg. fault and which
is machine dependent which means that the code's behavior
is undeterministic in different device. The main error is not in
the tested function but in the test suit, and the bug is caused
by the map container not been well initialized.