Fix the bug in lot_t::setCanRunLocation function
The bug is discovered when I want to accurately count the
scheduled jobs and I found that the number of unscheduled job
and scheduled jobs is not consistent. I trace the code carefully
and I found that in previous version, not all jobs are scheduled
in stage 2 scheduling process, the result is that the location
constraint is totally wrong. I fix the bug in the function.