Fix the bug caused by empty oper in IDLE machine
In previous version, the code attempt to create the lot_t instance
which is currently on the machine. However, the IDLE machine's
oper and qty are empty and which cause the string-to-integer
conversion error so that the machine would not be in the schedule
plan. In this version, I fix it by given 0 to replace empty string
of qty and oper.