Fix the bug caused by removing code in prev. ver.
In previous version, I removed the wrong code causing producing
wrong result. The removed code is in machines.cpp and in the
function scheduleGroups. The code is used to collect the
scheduled jobs and also set the last job to the machine. In prev.
ver. I remove the code to debug and test, but I totally forgot
to add the code back.