Fix the transportation time bug
In previous version, the transportation time is calculated by
two loations, one is from current job and the other is from
prev_job. I fix the bug changing the way of calculation of
current job's location and the location of machine.

The bug is fixed in the function scheduling in machine.cpp.