Fix the bug of setDAQueueTime
In previous version, the time difference in the function
setDAQueueTime is in second unit. I change the unit in
this version. Futhermore, there are some errors in input data,
which is the time format. I add the regex to check the time
format if is 20XX-XX-XX XX:XX or XX-XX-XX XX:XX. If the format
is former, the program change the format so that
the return value of timeConverter is correct.