Change some variable names for config
Previously, lot.h will find if "PXX" is substring of _lot_number,
after this commit, it now find if "XX" is the second and third
character of _lot_number.
As I am unsure if the program wants to match only the second and
third character, the find substring version of "XX" is left as
a comment above the current implementation

In main.cpp, the argument "OPER" is changed to "SETTING OPER" as
requested. Typo "process_tsime" to "process_time" is fixed as well.

In lot.cpp, there is a request to match _last_entity with "DA", and
if it is "DA" assign WB_move_in_entity to _last_entity. As the
request is currently not clear enough, the implementation currently
is wrapped as comment.