Add function object timeConverter
In this commit, I wrote a function object timeConverter
which is going to replace the old function declared in
include/infra.h used to transform the time format string to
a time stamp a.k.a seconds since JAN 01 1900.

In this commit ,I also wrote several test cases used to test
the function object timeConverter. Now, the time format string
can be in the formats listed below:
    YY-MM-DD HH:MM
    YY-MM-DD HH:MM:SS
    YYYY\MM\DD HH:MM:SS
Y means the digit of year. M means the digit of month.
D means the digit of day. H means the digit of hour.
S means the digit of second.

The test suit is written in
/tests/test_infrastructure/test_time_converter.cpp .