Initialize time zone in unit test
In previous version, time zone is not set in tests/main.cpp
so that the unit test failed in the github action because
the test suit use GMT+8 unix-time stamp for testing the
object timeConverter. However, the machine of github action
running the test may not be in the time zone GMT+8 so that
the test failed very often. In this version, I fix it.