Add new feature in timeConverter and fix some bugs
In the previous version, if the time digit is single the text
is unable to parse because of bug in regex string

In this commit I also make these kinds of string, "YYYY/MM/DD" and
"YY-MM-DD" be able to be parsed. In these cases, set the time
to zero. So, the strings are actually treated as "YYYY/MM/DD 00:00"
and "YY-MM-DD 00:00".