Add more test suites for testing route
Test suites is added and ready to do optimization on route_t for
locating the index of oper fast. The newly added test suites are
for testing about getting the oper's index on route in the program
and test if lot is in scheduling plan [WB - 7, WB]. I add few test
cases and Yahan would probably be in charge adding more test cases,
hopefully.

In this commit, I also fix a small bug in testing setting up
stations before wb station. The bug is in test cases. I added some
redundent stations when I was writting the test case so that the
test case loses its function.