Add test suite for testing setting up queue time
For the user who is not familiar with debugging, it is too difficult
to find out the faulty in the file. I add some failure messages
for the condition when the file is in invalid format or contains
some error data. The error mssage would point out the invalid
data in the queue_time file so that user can easily debug and
re-run the application.

The test suite also tests the error message when throwing out the
exception.