Optimize function findStationIdx
In this version, I optimize the function findStationIdx to be
safe and efficient. I was supposed to use branch-less technique
but I found that it isn't necessary. I just used bool operator to
finish it.

In this version, I also add some test cases in test_oper_index.cpp
for testing the function if the inexistent routename would become
existent after calling the function.