Rearrange lots.csv
In previous version, Kuka ask us to make sure that the field
CAN_RUN_MODELS should include all of models. We did it for her
and unfortunately, it results an error in rerun process. In
this commit, I fix it by adding another field for outputing all
models and remaining CAN_RUN_MODELS for the models which exactly
have precise process time.