Add some test suites for testing queuetime
In this version, I add a bunch of test suites for testing
calculating queue time. The process of verfying the function
calculate_queue_time is complcated. The test suites are used to
simulate current status of lot, therefore, the test suites contains
these information, route, process_id, mvin, is sublot or not,
current operation. The unit test writer should provide these exactly
information to calculate the queue time precisely. The remaining
fields in test case I haven't mantioned are the fields for expected
queue time and the status code in the route traverse. The return
value of route traverse is the or operation of several enum number.
It provides the return value carring more then one condition
information, so that after getting such return value, the remaining
part of code is able to deal with such condition. The enum flags
are TRAVERSE_DA_UNARRIVED, TRAVERSE_DA_ARRIVED, TRAVERSE_DA_MVIN
and TRAVERSE_FINISHED.